(add) DB connection function moved to js file
